Resolve integer overflow when running on a large 3D image
In a computation like x+y*nx+z*(nx*ny) if the sub-computation y*nx or
z*(nx*ny) overflows the size of a 32-bit integer, we will get a bogus
result even though the final destination is a 64-bit integer.
The solution is to promote at least one operand to a bigger-sized
integer. Since long can also be 32-bit on 32-bit architectures, we just
use explicit named types here.
This patch also fixes a few cases of comparisons of integers with
different signs.

Resolves: #235

+++ stardist/lib/stardist2d.cpp
@@ -550,7 +550,7 @@ static PyObject* c_non_max_suppression_inds(PyObject *self, PyObject *args) {
                                           results, params);
     else{
       results.resize(n_polys-i);
-      for (int n = 0; n < results.size(); ++n)
+      for (size_t n = 0; n < results.size(); ++n)
         results[n].first = i+n;
     }
     
@@ -563,7 +563,7 @@ static PyObject* c_non_max_suppression_inds(PyObject *self, PyObject *args) {
 #pragma omp parallel for schedule(dynamic) reduction(+:count_suppressed)   shared(suppressed)
 #endif
     
-    for (int neigh=0; neigh<results.size(); neigh++) {
+    for (size_t neigh=0; neigh<results.size(); neigh++) {
     // for (int j=i+1; j<n_polys; j++) {
 
       long j = results[neigh].first;
